move AggregateMethodType to measure package where it should have been..
